Move state types into reducers
