+++ src/app/selector/core.js
@@ -1,6 +1,9 @@
 // @flow
-import type {State} from 'App/type-next'
 
-export const selectMaterialGroups = (state: State) => state.core.materialGroups
-export const selectModels = (state: State) => state.core.models
-export const selectUploadingModels = (state: State) => state.core.uploadingModels
+import type {AppState, MaterialGroup, Model, UploadingModel} from 'App/type-next'
+
+export const selectMaterialGroups = (state: AppState): Array<MaterialGroup> =>
+  state.core.materialGroups
+export const selectModels = (state: AppState): Array<Model> => state.core.models
+export const selectUploadingModels = (state: AppState): Array<UploadingModel> =>
+  state.core.uploadingModels
