+++ src/app/type-next.js
@@ -1,6 +1,7 @@
 // @flow
 
 import type {AppAction as _AppAction} from './action-next'
+import type {AppState as _AppState} from './reducer-next'
 
 export type MaterialGroup = {
   id: string, // Generated by the frontend
@@ -99,33 +100,10 @@ export type ModalConfig = _ModalConfig<null | ModalContent>
 
 export type OpenModalConfig = _ModalConfig<ModalContent>
 
-export type UserState = {
-  userId: string | null,
-  location: Location | null,
-  currency: string
-}
-
-export type CoreState = {
-  models: Array<Model>,
-  uploadingModels: Array<UploadingModel>,
-  materialGroups: Array<MaterialGroup>
-}
-
-export type ModalState = ModalConfig & {
-  isOpen: boolean
-}
-
-export type State = {
-  core: CoreState,
-  modal: ModalState,
-  user: UserState,
-  routing: any, // Managed by react-router-redux
-  form: any // Managed by redux-form
-}
-
 export type Action<Type, Payload> = {
   type: Type,
   payload: Payload
 }
 
 export type AppAction = _AppAction
+export type AppState = _AppState
